[fix](auto inc) db_id and table_id should be int64_t instead of int32_t (#34912)

+++ be/src/olap/iterators.h
@@ -109,7 +109,7 @@ public:
     RuntimeState* runtime_state = nullptr;
     RowsetId rowset_id;
     Version version;
-    int32_t tablet_id = 0;
+    int64_t tablet_id = 0;
     size_t topn_limit = 0;
 };
